ENH: cluster/hierarchy: enabled Ward's algorithm for a distance matrix.
Previously, if this method (or centroid, or median) was selected the
linkage function only worked if the absolute positions of the datapoints
were provided (the docs didn't mention it, and I didn't update them so
they still don't). But for Ward's algorithm (and, actually, for
centroid and median as well) a distance function is more than enough,
provided that the metric we're working with is Euclidean.
Ward's algorithm existing implementation already allowed for this
kind of use, so I just needed to lift an ad-hoc restriction to
enable this feature.
The downside is that now we _assume_ the metric to be euclidean, but we
don't enforce it in any way. By comparison, the equivalent MATLAB
function issues a warning if it finds that the metric used may be
non-euclidean, while octave's documentation states that such an
assumption is always made.